+++ function.h
@@ -2,7 +2,7 @@
 #ifndef FUNCTION_H
 #define FUNCTION_H
 
-void unit_area_c(int tech,int celltype,int xbarsize,int adposition,int adderposition,int pulseposition,int sig_bit,int application,int rramtech,int read_seperate);
+void unit_area_c(int tech,int celltype,int xbarsize,int adposition,int adderposition,int pulseposition,int sig_bit,int application,double rramtech,int read_seperate);
 void unit_latency_c(int tech,int celltype,int xbarsize,int adposition,int adderposition,int pulseposition,int action_type,int sig_bit,int read_seperate);
 void unit_power_c(int tech,int celltype,int xbarsize,int adposition,int adderposition,int pulseposition,int action_type,int sig_bit,int application,int resis_range,int netrow,double xbar_latency,double adda_latency,double adder_latency,double decoder_latency,double write_latency,double read_latency,double read_seperate);
 void accuracy_c(int xbarsize,int linetech,int sig_bit,int cell_bit,int resis_range,int input_err);
